refact(#24): ViewHolder 외부 클래스로 수정
